+++ src/components/user/repo/index.js
@@ -1,9 +1,11 @@
 const userRepo = require('./user.repo');
 const accountRepo = require('./account.repo');
 const trashbinRepo = require('./trashbin.repo');
+const registrationPinRepo = require('./registrationPin.repo');
 
 module.exports = {
-	userRepo,
-	trashbinRepo,
 	accountRepo,
+	registrationPinRepo,
+	trashbinRepo,
+	userRepo,
 };