+++ test/services/helpdesk/index.test.js
@@ -1,5 +1,7 @@
+// eslint-disable no-process-env
 const assert = require('assert');
 const { expect } = require('chai');
+const sinon = require('sinon');
 const appPromise = require('../../../src/app');
 
 describe('helpdesk service', function test() {
@@ -7,6 +9,8 @@ describe('helpdesk service', function test() {
 	let app;
 	let helpdeskService;
 	let logger;
+	let originalMailService;;
+
 	const testProblem = {
 		type: 'contactAdmin',
 		_id: '5836bb5664582c35df3bc214',
@@ -16,14 +20,22 @@ describe('helpdesk service', function test() {
 		schoolId: '5836bb5664582c35df3bc000',
 	};
 
+	function MockMailService() {
+		return {
+			create: sinon.fake.returns(Promise.resolve()),
+		};
+	}
+
 	before(async () => {
 		app = await appPromise;
+		originalMailService = app.service('mails');
 		helpdeskService = app.service('helpdesk');
 		({ logger } = app);
 		await helpdeskService.create(testProblem);
 	});
 
 	after((done) => {
+		app.use('/mails', originalMailService);
 		helpdeskService
 			.remove(testProblem)
 			.then((result) => {
@@ -92,6 +104,54 @@ describe('helpdesk service', function test() {
 			expect(result).to.equal({});
 			expect(result.replyTo).to.equal('test@mail.de');
 		});
+	});
+
+	it('POST /helpdesk to schoolcloud with problem and without theme should pass proper email in argument', async () => {
+		const postBody = {
+			type: 'contactHPI',
+			supportType: 'problem',
+			subject: 'Dies ist ein Titel 4',
+			problemDescription: 'Dies ist die Problembeschreibung 1',
+			replyEmail: 'test@mail.de',
+		};
+		const mailService = new MockMailService();
+		app.use('/mails', mailService);
+		await helpdeskService.create(postBody, { account: { userId: '0000d213816abba584714c0a' } });
+		expect(mailService.create.firstArg.email).to.equal('ticketsystem@schul-cloud.org');
+	});
+
+	it('POST /helpdesk to schoolcloud with problem and with n21 theme should pass proper email in argument based on the supportType', async () => {
+		const postBody = {
+			type: 'contactHPI',
+			supportType: 'problem',
+			subject: 'Dies ist ein Titel 4',
+			problemDescription: 'Dies ist die Problembeschreibung 1',
+			replyEmail: 'test@mail.de',
+		};
+		const mailService = new MockMailService();
+		app.use('/mails', mailService);
+		const tempScTheme = process.env.SC_THEME;
+		process.env.SC_THEME = 'n21';
+		await helpdeskService.create(postBody, { account: { userId: '0000d213816abba584714c0a' } });
+		expect(mailService.create.firstArg.email).to.equal('nbc-support@netz-21.de');
+		process.env.SC_THEME = tempScTheme;
+	});
+
+	it('POST /helpdesk to schoolcloud with problem and with n21 theme should pass proper email in argument based on the supportType', async () => {
+		const postBody = {
+			type: 'contactHPI',
+			supportType: 'wish',
+			subject: 'Dies ist ein Titel 4',
+			problemDescription: 'Dies ist die Problembeschreibung 1',
+			replyEmail: 'test@mail.de',
+		};
+		const mailService = new MockMailService();
+		app.use('/mails', mailService);
+		const tempScTheme = process.env.SC_THEME;
+		process.env.SC_THEME = 'n21';
+		await helpdeskService.create(postBody, { account: { userId: '0000d213816abba584714c0a' } });
+		expect(mailService.create.firstArg.email).to.equal('nbc-wunsch@netz-21.de');
+		process.env.SC_THEME = tempScTheme;
 	});
 
 	it('POST /helpdesk to schoolcloud with feedback, valid data', () => {