+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 ## [25.2.3]
 
-### Added
+### Changed
 
 - SC-6727 Change email addresses for tickets for Niedersachsen